requirements.txt
aiogram==2.14.3
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
arrow==1.3.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
blessed==1.20.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.0.1
django-picklefield==3.2
environs==10.3.0
frozenlist==1.5.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gspread==6.2.0
gspread-dataframe==4.0.0
httplib2==0.22.0
idna==3.10
marshmallow==3.26.1
multidict==6.1.0
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==10.2.0
propcache==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==3.5.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
yarl==1.18.3
